using a script to build the config
